% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_tapestry_loaded}
\alias{check_tapestry_loaded}
\title{Check that Tapestry package has loaded successfully}
\usage{
check_tapestry_loaded()
}
\description{
Simple function to check that Tapestry package has loaded 
  successfully. Prints "Tapestry loaded successfully!" if so.
}
