% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestry_project.R
\name{is.tapestry_project}
\alias{is.tapestry_project}
\title{Determine if object is of class tapestry_project}
\usage{
is.tapestry_project(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Determine if object is of class tapestry_project.
}
\details{
TODO
}
\examples{
# TODO

}
