% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestry_project.R
\name{print.tapestry_project}
\alias{print.tapestry_project}
\title{Custom print function for class tapestry_project}
\usage{
\method{print}{tapestry_project}(x, ...)
}
\arguments{
\item{x}{object of class \code{tapestry_project}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class tapestry_project, printing a 
  summary of the key elements (also equivalent to \code{summary(x)}). To do 
  an ordinary \code{print()} of all elements of the project, use the 
  \code{print_full()} function.
}
