% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Test function to run an example MCMC}
\usage{
run_mcmc(
  a,
  r,
  p,
  burnin = 100,
  samples = 1000,
  beta = 1,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{a}{vector of alternative allele counts.}

\item{r}{vector of reference allele counts.}

\item{p}{vector of allele frequencies per locus.}

\item{burnin}{the number of burn-in iterations.}

\item{samples}{the number of sampling iterations.}

\item{beta}{vector of thermodynamic powers. Final value in the vector should
always be 1.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100\%. This avoids large amounts of
output being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
This should be replaced by a more carefully thought out
  structure, probably involving defining a parameters dataframe outside of
  this function that can be loaded in.
}
