% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestry_project.R
\name{summary.tapestry_project}
\alias{summary.tapestry_project}
\title{Print summary for class tapestry_project}
\usage{
\method{summary}{tapestry_project}(object, ...)
}
\arguments{
\item{object}{object of class \code{tapestry_project}}

\item{...}{other arguments (ignored)}
}
\description{
Overload summary function for class tapestry_project
}
