% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestry_project.R
\name{tapestry_project}
\alias{tapestry_project}
\title{Define empty tapestry_project object}
\usage{
tapestry_project()
}
\description{
Define empty tapestry_project object
}
\details{
TODO
}
\examples{
# TODO

}
